requirements.txt
flask>=3.0.0
flask-jwt-extended>=4.6.0
werkzeug>=3.0.0
authlib>=1.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
psycopg>=3.1.0           
pgvector>=0.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-groq>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
openai>=1.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
langfuse>=2.0.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
yfinance>=0.2.0
vnstock>=0.2.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
ragas>=0.1.0
datasets>=2.0.0
